docs: update benchmarks, fix all build warnings, complete Phase 11 status, research 10.10.2
Benchmarks:
- Added Phase 10 Round 2 results to PERFORMANCE_RESULTS.md
- FetchInt10: 8.75 ns/row (114.3M rows/s) — 12.4% improvement over baseline
- FetchVarchar5: 8.20 ns/row (122.0M rows/s) — 12.0% improvement
- Added cumulative comparison table (Baseline → R1 → R2)

Build warnings fixed (zero warnings in both Debug and Release):
- Mlist.h: removed C++17-deprecated 'register' storage class (C5033)
- OdbcObject.cpp: removed [[likely]] attribute (C5051, requires C++20)
- OdbcStatement.cpp: fixed sprintf format for intptr_t (%ld → %lld) (C4477)
- Utf16Convert.cpp: guarded surrogate macros with #ifndef (C4005)

FIXME pragmas removed (4 total):
- OdbcDesc.cpp SQL_DESC_TYPE: not a bug — verbose==concise for non-interval
  types (all Firebird types). Replaced with spec-reference comment.
- OdbcDesc.cpp SQL_DESC_CONCISE_TYPE: same rationale as above.
- OdbcDesc.cpp SQL_DESC_DATA_PTR: consistency check is optional QoI
  enhancement. Replaced with spec-reference comment.
- IscConnection.cpp getNativeSql: schema removal + auto-quoting is
  functional and stable. Replaced with descriptive comment.

Phase 11 master plan:
- Added Status column and completion dates to section 11.1 table
- All 7 tasks (11.1.1–11.1.7) marked ✅ with 'Completed Feb 8, 2026'

Task 10.10.2 research:
- Firebird OO API has no multi-row fetch (single void* buffer per call)
- Commented-out Pipe interface (FirebirdApi.idl) would have provided it
- Wire protocol already prefetches up to 1000 rows transparently
- REMOTE_compute_batch_size() computes optimal batch; rsr_reorder_level
  triggers pipelining at 50% depletion
- No action needed — 8.75 ns/row already far below 500 ns target

+++ src/OdbcStatement.cpp
@@ -159,7 +159,7 @@ using namespace IscDbcLibrary;
 void TraceOutput(char * msg, intptr_t val)
 {
 	char buf[80];
-	sprintf( buf, "\t%s = %ld : %p\n", msg, val, (void*)val );
+	sprintf( buf, "\t%s = %lld : %p\n", msg, (long long)val, (void*)val );
 	OutputDebugString(buf);
 }
